requirements.txt
email-validator>=2.3.0
flask-cors>=6.0.1
flask>=3.1.2
flask-sqlalchemy>=3.1.1
google-genai>=1.32.0
gunicorn>=23.0.0
openai>=1.102.0
psycopg2-binary>=2.9.10
trafilatura>=2.0.0
werkzeug>=3.1.3
pypdf2>=3.0.1
groq>=0.31.0
requests>=2.32.5
flask-login>=0.6.3
sqlalchemy>=2.0.43
PyPDF2
langchain
google-generativeai
langchain-google-genai
faiss-cpu
python-dotenv